requirements.txt
flask==2.3.3
flask-login==0.6.2
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
langchain==0.1.12
langchain-community==0.0.27
langchain-core==0.1.30
langgraph==0.0.27
langchain-openai==0.0.8
langchain-google-genai==0.0.7
langchain-anthropic==0.1.4
python-dotenv==1.0.1
requests==2.31.0
sqlalchemy==2.0.27
pydantic==2.6.1
chromadb==0.4.22
pymongo==4.6.1
gunicorn==21.2.0
pytest==7.4.3
